Moved asc io ui and property definitions
