+++ Arma3ObjectBuilder/ui/__init__.py
@@ -1,4 +1,3 @@
-from . import import_export_asc
 from . import import_export_mcfg
 from . import import_export_p3d
 from . import import_export_rtm
